Updated Profile Button and other changes
