+++ libs/ui/src/components/atoms/Button/index.ts
@@ -1,1 +1,2 @@
 export * from './Button';
+export type { ButtonProps } from './Button.types';
