--- libs/ui/src/components/molecules/DropdownButton/index.ts
@@ -1,1 +1,0 @@
-export * from './DropdownButton';
