+++ libs/ui/src/components/molecules/ProfileButton/ProfileButton.types.ts
@@ -1,0 +1,8 @@
+import { ButtonProps } from '../../atoms/Button/Button.types';
+import { ProfileAvatarProps } from '../ProfileAvatar/ProfileAvatar';
+
+type OmittedProps = 'IconLeft';
+export interface ProfileButtonProps extends Omit<ButtonProps, OmittedProps> {
+  /* Profile Avatar stils on the left of button */
+  profile: Omit<ProfileAvatarProps, 'size'>;
+}
